FLW-733 Payouts list: implement viewModel factory
